fix survey questions

+++ backend/som/api/survey/create_survey.py
@@ -1,5 +1,6 @@
 import os
 from pathlib import Path
+import traceback
 from sqlalchemy.orm import Session
 from schemas.survey_number_schema import SurveyNumberResponse
 import shutil
@@ -26,6 +27,8 @@ def save_files(files: list[UploadFile], file_type: str, survey_number: int):
     os.makedirs(save_dir, exist_ok=True)
 
     for file in files:
+        if not file:
+            continue
         file_path = save_dir / file.filename
         with open(file_path, "wb") as buffer:
             shutil.copyfileobj(file.file, buffer)
@@ -36,6 +39,8 @@ def delete_files(files: list[UploadFile], file_type: str, survey_number: int):
         return
     
     for file in files:
+        if not file:
+            continue
         file_path = UPLOAD_DIR / str(survey_number) / file_type / file.filename
         if file_path.exists():
             file_path.unlink()
@@ -65,22 +70,24 @@ def create_survey(
     db.flush()
     db.refresh(survey_number)
 
-    create_survey_questions(survey_number_id=survey_number.id, questions_file=questions_file, db=db)
-    create_survey_hunt(survey_number_id=survey_number.id, hunt_file=hunt_file, db=db)
-    create_survey_files(survey_number_id=survey_number.id, file_extension="csv", files=csv_files, db=db)
-    create_survey_files(survey_number_id=survey_number.id, file_extension="ecg", files=ecg_files, db=db)
-    create_survey_files(survey_number_id=survey_number.id, file_extension="hr", files=hr_files, db=db)
-    create_survey_files(survey_number_id=survey_number.id, file_extension="iqdat", files=iqdat_files, db=db)
-    create_survey_files(survey_number_id=survey_number.id, file_extension="mp4", files=mp4_files, db=db)
-    create_survey_files(survey_number_id=survey_number.id, file_extension="rr", files=rr_files, db=db)
-    create_survey_files(survey_number_id=survey_number.id, file_extension="sm", files=sm_files, db=db)
-    create_survey_files(survey_number_id=survey_number.id, file_extension="tmk", files=tmk_files, db=db)
-    create_survey_files(survey_number_id=survey_number.id, file_extension="txt", files=txt_files, db=db)
-    create_survey_files(survey_number_id=survey_number.id, file_extension="other", files=other_files, db=db)
-
     try:
-        save_files(questions_file, "questions", survey_number=survey_number.number)
-        save_files(hunt_file, "hunt", survey_number=survey_number.number)
+        create_survey_questions(survey_number_id=survey_number.id, questions_file=questions_file, db=db)
+        print("aaaaaaaa")
+        create_survey_hunt(survey_number_id=survey_number.id, hunt_file=hunt_file, db=db)
+        print("aaaaaaaa")
+        create_survey_files(survey_number_id=survey_number.id, file_extension="csv", files=csv_files, db=db)
+        create_survey_files(survey_number_id=survey_number.id, file_extension="ecg", files=ecg_files, db=db)
+        create_survey_files(survey_number_id=survey_number.id, file_extension="hr", files=hr_files, db=db)
+        create_survey_files(survey_number_id=survey_number.id, file_extension="iqdat", files=iqdat_files, db=db)
+        create_survey_files(survey_number_id=survey_number.id, file_extension="mp4", files=mp4_files, db=db)
+        create_survey_files(survey_number_id=survey_number.id, file_extension="rr", files=rr_files, db=db)
+        create_survey_files(survey_number_id=survey_number.id, file_extension="sm", files=sm_files, db=db)
+        create_survey_files(survey_number_id=survey_number.id, file_extension="tmk", files=tmk_files, db=db)
+        create_survey_files(survey_number_id=survey_number.id, file_extension="txt", files=txt_files, db=db)
+        create_survey_files(survey_number_id=survey_number.id, file_extension="other", files=other_files, db=db)
+        print("aaaaaaaa")
+        save_files([questions_file], "questions", survey_number=survey_number.number)
+        save_files([hunt_file], "hunt", survey_number=survey_number.number)
         save_files(csv_files, "csv", survey_number=survey_number.number)
         save_files(ecg_files, "ecg", survey_number=survey_number.number)
         save_files(hr_files, "hr", survey_number=survey_number.number)
@@ -92,8 +99,9 @@ def create_survey(
         save_files(txt_files, "txt", survey_number=survey_number.number)
         save_files(other_files, "other", survey_number=survey_number.number)
     except Exception as e:
-        delete_files(questions_file, "questions", survey_number=survey_number.number)
-        delete_files(hunt_file, "hunt", survey_number=survey_number.number)
+        traceback.print_exc()
+        delete_files([questions_file], "questions", survey_number=survey_number.number)
+        delete_files([hunt_file], "hunt", survey_number=survey_number.number)
         delete_files(csv_files, "csv", survey_number=survey_number.number)
         delete_files(ecg_files, "ecg", survey_number=survey_number.number)
         delete_files(hr_files, "hr", survey_number=survey_number.number)
